+++ src/main/java/Lilith.java
@@ -1,8 +1,4 @@
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.*;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
 
 public class Lilith {
@@ -164,204 +160,5 @@ public class Lilith {
     }
 }
 
-class Task{
-    enum TaskType{
-        ToDos, Deadline, Events
-    }
 
-    String taskname;
-    boolean status;
-    TaskType tasktype;
-    String startdetail;
-    String enddetail;
-
-    Task(String taskname, String startdetail, String enddetail){
-        this.taskname = taskname;
-        this.status = false;
-        this.tasktype = TaskType.ToDos;
-        this.startdetail = startdetail;
-        this.enddetail = enddetail;
-    }
-
-    public void setTask(TaskType input){
-        tasktype = input;
-    }
-
-    void mark(){
-        status = true;
-    }
-
-    void unmark(){
-        status = false;
-    }
-
-    public String toFileString() {
-        String typeLetter = switch (tasktype) {
-            case ToDos -> "T";
-            case Deadline -> "D";
-            case Events -> "E";
-        };
-
-        String done = status ? "1" : "0";
-
-        if (tasktype == TaskType.ToDos) {
-            return typeLetter + " | " + done + " | " + taskname;
-        }
-
-        if (tasktype == TaskType.Deadline) {
-            return typeLetter + " | " + done + " | " + taskname + " | " + enddetail;
-        }
-
-        // Events
-        return typeLetter + " | " + done + " | " + taskname
-                + " | " + startdetail + " | " + enddetail;
-    }
-
-    public static Task fromFileString(String line) {
-        String[] parts = line.split("\\s*\\|\\s*");
-
-        if (parts.length < 3) {
-            throw new IllegalArgumentException("Invalid task format");
-        }
-
-        String type = parts[0];
-        boolean done = parts[1].equals("1");
-        String name = parts[2];
-        Task task;
-
-        switch (type) {
-            case "T" -> {
-                task = new Task(name, null, null);
-                task.setTask(TaskType.ToDos);
-            }
-
-            case "D" -> {
-                if (parts.length < 4) {
-                    throw new IllegalArgumentException("Deadline missing detail");
-                }
-                task = new Task(name, null, parts[3]);
-                task.setTask(TaskType.Deadline);
-            }
-
-            case "E" -> {
-                if (parts.length < 5) {
-                    throw new IllegalArgumentException("Event missing details");
-                }
-                task = new Task(name, parts[3], parts[4]);
-                task.setTask(TaskType.Events);
-            }
-
-            default -> throw new IllegalArgumentException("Unknown task type");
-        }
-
-        if (done) {
-            task.mark();
-        }
-
-        return task;
-    }
-
-
-    @Override
-    public String toString() {
-        switch (tasktype){
-            case ToDos -> {
-                return "[T][" + (status ? "X" : " ") + "] " + taskname;
-            }
-            case Deadline -> {
-                return "[D][" + (status ? "X" : " ") + "] " + taskname + "(by: " + enddetail + ")";
-            }
-            case Events -> {
-                return "[E][" + (status ? "X" : " ") + "] " + taskname + "(from: " + startdetail + " to: " + enddetail + ")";
-            }
-        }
-        return "[" + (status ? "X" : " ") + "] " + taskname;
-    }
-} 
-
-
-class Storage {
-
-    private final Path filePath;
-
-    public Storage(String path) {
-        this.filePath = Paths.get(path);
-    }
-
-    /**
-     * Ensures that the folder and file exist.
-     * Creates them if missing.
-     */
-    private void ensureFileExists() throws IOException {
-
-        // Create ./data folder if missing
-        if (!Files.exists(filePath.getParent())) {
-            Files.createDirectories(filePath.getParent());
-        }
-
-        // Create file if missing
-        if (!Files.exists(filePath)) {
-            Files.createFile(filePath);
-        }
-    }
-
-    /**
-     * Loads tasks from the file.
-     * Skips corrupted lines safely.
-     */
-    public ArrayList<Task> loadTasks() {
-
-        ArrayList<Task> tasks = new ArrayList<>();
-
-        try {
-            ensureFileExists();
-
-            // Read all lines at once
-            List<String> lines = Files.readAllLines(filePath, StandardCharsets.UTF_8);
-
-            for (String line : lines) {
-
-                if (line.trim().isEmpty()) {
-                    continue;
-                }
-
-                try {
-                    Task task = Task.fromFileString(line);
-                    tasks.add(task);
-                } catch (Exception e) {
-                    System.out.println("Skipping corrupted line: " + line);
-                }
-            }
-
-        } catch (IOException e) {
-            System.out.println("Error loading tasks: " + e.getMessage());
-        }
-
-        return tasks;
-    }
-
-    /**
-     * Saves tasks into the file using Files.write().
-     * Overwrites the file each time.
-     */
-    public void saveTasks(ArrayList<Task> tasks) {
-
-        try {
-            ensureFileExists();
-
-            List<String> lines = new ArrayList<>();
-
-            // Convert each task into a file line
-            for (Task task : tasks) {
-                lines.add(task.toFileString());
-            }
-
-            // Write all lines into file (overwrite)
-            Files.write(filePath, lines, StandardCharsets.UTF_8);
-
-        } catch (IOException e) {
-            System.out.println("Error saving tasks: " + e.getMessage());
-        }
-    }
-}
 //test branching
